move service tests into service directory
